refactor : BigInteger -> Long 수정

+++ boottalk/src/main/java/com/icandoit/boottalk/user/controller/ManageController.java
@@ -3,7 +3,6 @@ package com.icandoit.boottalk.user.controller;
 import com.icandoit.boottalk.common.dto.BaseResponse;
 import com.icandoit.boottalk.user.domain.form.UpdateForm;
 import com.icandoit.boottalk.user.service.ManageService;
-import java.math.BigInteger;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -19,28 +18,29 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/users/my")
 public class ManageController {
 
-  private final ManageService manageService;
+	private final ManageService manageService;
 
-  @GetMapping
-  public ResponseEntity<BaseResponse> getUser(@RequestParam BigInteger userId) {
+	@GetMapping
+	public ResponseEntity<BaseResponse> getUser(@RequestParam Long userId) {
 
-    return ResponseEntity.ok(new BaseResponse("회원정보가 조회되었습니다.", manageService.getUser(userId)));
-  }
+		return ResponseEntity.ok(new BaseResponse("회원정보가 조회되었습니다.", manageService.getUser(userId)));
+	}
 
-  @PutMapping
-  public ResponseEntity<BaseResponse> updateUser(@RequestParam BigInteger userId, @RequestBody UpdateForm form) {
+	@PutMapping
+	public ResponseEntity<BaseResponse> updateUser(@RequestParam Long userId,
+		@RequestBody UpdateForm form) {
 
-    return ResponseEntity.ok(new BaseResponse("회원정보가 수정되었습니다.", manageService.updateUser(userId, form)));
-  }
+		return ResponseEntity.ok(
+			new BaseResponse("회원정보가 수정되었습니다.", manageService.updateUser(userId, form)));
+	}
 
-  @DeleteMapping
-  public ResponseEntity<BaseResponse> deleteUser(@RequestParam BigInteger userId) {
+	@DeleteMapping
+	public ResponseEntity<BaseResponse> deleteUser(@RequestParam Long userId) {
 
-    manageService.deleteUser(userId);
+		manageService.deleteUser(userId);
 
-    return ResponseEntity.ok(new BaseResponse("회원 탈퇴되었습니다.", null));
-  }
-
+		return ResponseEntity.ok(new BaseResponse("회원 탈퇴되었습니다.", null));
+	}
 
 
 }
